Write Higher Order Arrow Functions
It's time we see how powerful arrow functions are when processing data.

Arrow functions work really well with higher order functions, such as map(), filter(), and reduce(), that take other functions as arguments for processing collections of data.

Read the following code:

FBPosts.filter(function(post) {
  return post.thumbnail !== null && post.shares > 100 && post.likes > 500;
})
We have written this with filter() to at least make it somewhat readable. Now compare it to the following code which uses arrow function syntax instead:

FBPosts.filter((post) => post.thumbnail !== null && post.shares > 100 && post.likes > 500)
This code is more succinct and accomplishes the same task with fewer lines of code.


Use arrow function syntax to compute the square of only the positive integers (decimal numbers are not integers) in the array realNumberArray and store the new array in the variable squaredIntegers. 

PS: Instead of using "parseInt(i)" to make the function look for the integers that can be divided by the numbers that remained after applying the first filter ("i > 0"), I tried using "1" because it gives the same outcome (after filtering "i >0"), since any positive number within the array that can be divided by 1, and has a remainder of 0 ("i % 1 === 0") is necessarily an integer. The final part of the code adds the ".map()" filter function to find the remaining integers (after the ".filter()") within the array to square them with ("Math.pow (i, x)" i being a base integer from the array, and x being the exponent).